test-requirements.txt
fakeredis
mock
nose >= 1.3.0
nose-cov
